+++ 2004/taeseongyun/data/src/main/java/com/tsdev/data/model/MovieResponse.kt
@@ -1,0 +1,11 @@
+package com.tsdev.data.model
+
+import com.tsdev.domain.model.Item
+
+data class MovieResponse(
+    val display: Int,
+    val items: List<Item>,
+    val lastBuildDate: String,
+    val start: Int,
+    val total: Int
+)
