+++ 2004/taeseongyun/domain/src/main/java/com/tsdev/domain/di/MapperModule.kt
@@ -1,9 +1,9 @@
 package com.tsdev.domain.di
 
-import com.tsdev.data.source.Item
-import com.tsdev.data.source.MovieResponse
 import com.tsdev.domain.mapper.Mapper
 import com.tsdev.domain.mapper.MapperImpl
+import com.tsdev.domain.model.Item
+import com.tsdev.domain.model.MovieResponse
 import org.koin.dsl.module
 
 val domainMapperModule = module {
