--- 2004/taeseongyun/domain/src/main/java/com/tsdev/domain/di/RepositoryModule.kt
@@ -1,9 +1,0 @@
-package com.tsdev.domain.di
-
-import com.tsdev.domain.repository.NaverMovieRepository
-import com.tsdev.domain.repository.NaverMovieRepositoryImpl
-import org.koin.dsl.module
-
-val repositoryModule = module {
-    single<NaverMovieRepository> { NaverMovieRepositoryImpl(get(), get()) }
-}
