+++ 2004/taeseongyun/domain/src/main/java/com/tsdev/domain/model/Item.kt
@@ -1,12 +1,4 @@
-package com.tsdev.data.source
-
-data class MovieResponse(
-    val display: Int,
-    val items: List<Item>,
-    val lastBuildDate: String,
-    val start: Int,
-    val total: Int
-)
+package com.tsdev.domain.model
 
 data class Item(
     val actor: String,
