+++ 2004/taeseongyun/domain/src/main/java/com/tsdev/domain/model/MovieResponse.kt
@@ -1,0 +1,9 @@
+package com.tsdev.domain.model
+
+data class MovieResponse(
+    val display: Int,
+    val items: List<Item>,
+    val lastBuildDate: String,
+    val start: Int,
+    val total: Int
+)
