+++ 2004/taeseongyun/presentation/build.gradle
@@ -46,5 +46,5 @@ dependencies {
     implementation project(':domain')
 
     //data
-    implementation project(':data')
+//    implementation project(':data')
 }
