+++ 2004/taeseongyun/remote/src/main/java/com/tsdev/remote/di/RemoteModule.kt
@@ -1,6 +1,6 @@
 package com.tsdev.remote.di
 
-import com.tsdev.data.source.remote.NaverMovieRemoteSourceData
+import com.tsdev.data.source.NaverMovieRemoteSourceData
 import com.tsdev.remote.NaverMovieRemoteDataSourceImpl
 import org.koin.dsl.module
