+++ 2004/taeseongyun/settings.gradle
@@ -1,6 +1,6 @@
 rootProject.name='My Application'
 include ':app'
-include ':data'
 include ':domain'
 include ':remote'
 include ':presentation'
+include ':data'
